Gemini Health Mate/requirements.txt
streamlit
requests
Pillow
SpeechRecognition
pyaudio
streamlit
google.generativeai
elevenlabs